+++ src/base/widget.c
@@ -44,10 +44,10 @@ ret_t widget_move(widget_t* widget, xy_t x, xy_t y) {
   if (widget->x != x || widget->y != y) {
     widget_dispatch(widget, &e);
 
-    widget_invalidate_force(widget);
+    widget_invalidate_force(widget, NULL);
     widget->x = x;
     widget->y = y;
-    widget_invalidate_force(widget);
+    widget_invalidate_force(widget, NULL);
 
     e.type = EVT_MOVE;
     widget_dispatch(widget, &e);
@@ -63,10 +63,10 @@ ret_t widget_resize(widget_t* widget, wh_t w, wh_t h) {
   if (widget->w != w || widget->h != h) {
     widget_dispatch(widget, &e);
 
-    widget_invalidate_force(widget);
+    widget_invalidate_force(widget, NULL);
     widget->w = w;
     widget->h = h;
-    widget_invalidate_force(widget);
+    widget_invalidate_force(widget, NULL);
 
     e.type = EVT_RESIZE;
     widget_dispatch(widget, &e);
@@ -82,12 +82,12 @@ ret_t widget_move_resize(widget_t* widget, xy_t x, xy_t y, wh_t w, wh_t h) {
   if (widget->x != x || widget->y != y || widget->w != w || widget->h != h) {
     widget_dispatch(widget, &e);
 
-    widget_invalidate_force(widget);
+    widget_invalidate_force(widget, NULL);
     widget->x = x;
     widget->y = y;
     widget->w = w;
     widget->h = h;
-    widget_invalidate_force(widget);
+    widget_invalidate_force(widget, NULL);
 
     e.type = EVT_MOVE_RESIZE;
     widget_dispatch(widget, &e);
@@ -285,10 +285,10 @@ ret_t widget_set_state(widget_t* widget, widget_state_t state) {
   return_value_if_fail(widget != NULL, RET_BAD_PARAMS);
 
   if (widget->state != state) {
-    widget_invalidate_force(widget);
+    widget_invalidate_force(widget, NULL);
     widget->state = state;
     widget_update_style(widget);
-    widget_invalidate_force(widget);
+    widget_invalidate_force(widget, NULL);
   }
 
   return RET_OK;
@@ -337,7 +337,7 @@ ret_t widget_destroy_children(widget_t* widget) {
 }
 
 ret_t widget_add_child(widget_t* widget, widget_t* child) {
-  return_value_if_fail(widget != NULL && child != NULL, RET_BAD_PARAMS);
+  return_value_if_fail(widget != NULL && child != NULL && child->parent == NULL, RET_BAD_PARAMS);
 
   child->parent = widget;
   if (widget->children == NULL) {
@@ -356,6 +356,7 @@ ret_t widget_add_child(widget_t* widget, widget_t* child) {
 ret_t widget_remove_child(widget_t* widget, widget_t* child) {
   return_value_if_fail(widget != NULL && child != NULL, RET_BAD_PARAMS);
 
+  widget_invalidate_force(child, NULL);
   if (widget->target == child) {
     widget->target = NULL;
   }
@@ -376,6 +377,47 @@ ret_t widget_remove_child(widget_t* widget, widget_t* child) {
 
   child->parent = NULL;
   return array_remove(widget->children, NULL, child, NULL);
+}
+
+ret_t widget_insert_child(widget_t* widget, uint32_t index, widget_t* child) {
+  return_value_if_fail(widget != NULL && child != NULL, RET_BAD_PARAMS);
+  return_value_if_fail(widget_add_child(widget, child) == RET_OK, RET_FAIL);
+
+  return widget_restack(child, index);
+}
+
+ret_t widget_restack(widget_t* widget, uint32_t index) {
+  uint32_t i = 0;
+  uint32_t nr = 0;
+  uint32_t old_index = 0;
+  widget_t** children = NULL;
+  return_value_if_fail(widget != NULL && widget->parent != NULL, RET_BAD_PARAMS);
+
+  old_index = widget_index_of(widget);
+  nr = widget_count_children(widget->parent);
+  return_value_if_fail(old_index >= 0 && nr > 0, RET_BAD_PARAMS);
+
+  if (index >= nr) {
+    index = nr - 1;
+  }
+
+  if (index == old_index || nr == 1) {
+    return RET_OK;
+  }
+
+  children = (widget_t**)(widget->parent->children->elms);
+  if (index < old_index) {
+    for (i = old_index; i > index; i--) {
+      children[i] = children[i - 1];
+    }
+  } else {
+    for (i = old_index; i < index; i++) {
+      children[i] = children[i + 1];
+    }
+  }
+  children[index] = widget;
+
+  return RET_OK;
 }
 
 static widget_t* widget_lookup_child(widget_t* widget, const char* name) {
@@ -461,10 +503,10 @@ static ret_t widget_set_visible_self(widget_t* widget, bool_t visible) {
   return_value_if_fail(widget != NULL, RET_BAD_PARAMS);
 
   if (widget->visible != visible) {
-    widget_invalidate_force(widget);
+    widget_invalidate_force(widget, NULL);
     widget->visible = visible;
     widget_update_style(widget);
-    widget_invalidate_force(widget);
+    widget_invalidate_force(widget, NULL);
   }
 
   return RET_OK;
@@ -742,12 +784,12 @@ ret_t widget_set_prop(widget_t* widget, const char* name, const value_t* v) {
   } else if (tk_str_eq(name, WIDGET_PROP_H)) {
     widget->h = (wh_t)value_int(v);
   } else if (tk_str_eq(name, WIDGET_PROP_VISIBLE)) {
-    widget->visible = !!value_int(v);
+    widget->visible = value_bool(v);
   } else if (tk_str_eq(name, WIDGET_PROP_STYLE)) {
     const char* name = value_str(v);
     return widget_use_style(widget, name);
   } else if (tk_str_eq(name, WIDGET_PROP_ENABLE)) {
-    widget->enable = !!value_int(v);
+    widget->enable = value_bool(v);
   } else if (tk_str_eq(name, WIDGET_PROP_NAME)) {
     widget_set_name(widget, value_str(v));
   } else if (tk_str_eq(name, WIDGET_PROP_TEXT)) {
@@ -1211,7 +1253,7 @@ widget_t* widget_get_window(widget_t* widget) {
   return_value_if_fail(widget != NULL, NULL);
 
   while (iter) {
-    if (iter->is_window) {
+    if (iter->vt->is_window) {
       return iter;
     }
     iter = iter->parent;
@@ -1333,9 +1375,17 @@ ret_t widget_invalidate(widget_t* widget, rect_t* r) {
   }
 }
 
-ret_t widget_invalidate_force(widget_t* widget) {
+ret_t widget_invalidate_force(widget_t* widget, rect_t* r) {
+  widget_t* iter = widget;
   return_value_if_fail(widget != NULL, RET_BAD_PARAMS);
-  widget->dirty = FALSE;
+
+  while (iter != NULL) {
+    iter->dirty = FALSE;
+    if (iter->vt->is_window) {
+      break;
+    }
+    iter = iter->parent;
+  }
 
   return widget_invalidate(widget, NULL);
 }
@@ -1388,27 +1438,23 @@ widget_t* widget_init(widget_t* widget, widget_t* parent, const widget_vtable_t*
   if (parent != NULL && widget_is_window_opened(widget)) {
     widget_update_style(widget);
   }
-  widget_invalidate_force(widget);
+  widget_invalidate_force(widget, NULL);
 
   return widget;
 }
 
 ret_t widget_to_screen(widget_t* widget, point_t* p) {
-  value_t v;
   widget_t* iter = widget;
   return_value_if_fail(widget != NULL && p != NULL, RET_BAD_PARAMS);
 
   while (iter != NULL) {
+    if (iter->vt->scrollable) {
+      p->x -= widget_get_prop_int(iter, WIDGET_PROP_XOFFSET, 0);
+      p->y -= widget_get_prop_int(iter, WIDGET_PROP_YOFFSET, 0);
+    }
+
     p->x += iter->x;
     p->y += iter->y;
-
-    if (widget_get_prop(iter, WIDGET_PROP_XOFFSET, &v) == RET_OK) {
-      p->x -= value_int(&v);
-    }
-
-    if (widget_get_prop(iter, WIDGET_PROP_YOFFSET, &v) == RET_OK) {
-      p->y -= value_int(&v);
-    }
 
     iter = iter->parent;
   }
@@ -1421,6 +1467,11 @@ ret_t widget_to_local(widget_t* widget, point_t* p) {
   return_value_if_fail(widget != NULL && p != NULL, RET_BAD_PARAMS);
 
   while (iter != NULL) {
+    if (iter->vt->scrollable) {
+      p->x += widget_get_prop_int(iter, WIDGET_PROP_XOFFSET, 0);
+      p->y += widget_get_prop_int(iter, WIDGET_PROP_YOFFSET, 0);
+    }
+
     p->x -= iter->x;
     p->y -= iter->y;