+++ src/ext_widgets/scroll_view/list_item.c
@@ -59,11 +59,12 @@ static ret_t list_item_on_event(widget_t* widget, event_t* e) {
       list_item->down.x = evt->x;
       list_item->down.y = evt->y;
       list_item->timer_id = timer_add(list_item_on_timer, widget, 30);
-      widget_invalidate_force(widget);
+      widget_invalidate_force(widget, NULL);
       break;
     }
     case EVT_POINTER_DOWN_ABORT: {
       list_item_remove_timer(widget);
+      widget_invalidate_force(widget, NULL);
       widget_set_state(widget, WIDGET_STATE_NORMAL);
       break;
     }
@@ -76,7 +77,7 @@ static ret_t list_item_on_event(widget_t* widget, event_t* e) {
 
       list_item->dragged = FALSE;
       list_item_remove_timer(widget);
-      widget_invalidate_force(widget);
+      widget_invalidate_force(widget, NULL);
       widget_set_state(widget, WIDGET_STATE_NORMAL);
       break;
     }
@@ -84,7 +85,7 @@ static ret_t list_item_on_event(widget_t* widget, event_t* e) {
       pointer_event_t* evt = (pointer_event_t*)e;
       uint32_t dy = tk_abs(evt->y - list_item->down.y);
 
-      if (dy > 6) {
+      if (evt->pressed && dy > TK_DRAG_THRESHOLD) {
         list_item->dragged = TRUE;
         list_item_remove_timer(widget);
         widget_set_state(widget, WIDGET_STATE_NORMAL);